Rename ConnectionConfigReaderHandlers to ConnetionConfigReaders
